refactor: keep yml config clean from customize requirements
