+++ publiccode-parser/projects/civic-tech-experimental-field-dengue-backend.yml
@@ -54,6 +54,5 @@ tw:
     - name: "JavaScript"
   vulnerabilityScanners:
     - "OWASP"
-  mainCopyrightOwnerLogo: "/images/projects/logos/MODA-logo.svg"
   # GitHub API 有資訊可以取得, 暫時手動填寫
   createdDate: "2023-11-22"
