+++ publiccode-parser/projects/civic-tech-experimental-field-find-babycare-backend.yml
@@ -55,6 +55,5 @@ tw:
     - name: "Laravel"
   vulnerabilityScanners:
     - "OWASP"
-  mainCopyrightOwnerLogo: "/images/projects/logos/MODA-logo.svg"
   # GitHub API 有資訊可以取得, 暫時手動填寫
   createdDate: "2023-11-22"
