+++ publiccode-parser/projects/code-gov-tw.yml
@@ -48,6 +48,5 @@ tw:
     - "OWASP ZAP"
     - "SonarQube"
   accessibilityConformance: "AAA"
-  mainCopyrightOwnerLogo: "/images/projects/logos/MODA-logo.svg"
   # GitHub API 有資訊可以取得, 暫時手動填寫
   createdDate: "2024-02-01"
