+++ publiccode-parser/projects/gov-uk-forms-product-page.yml
@@ -45,9 +45,5 @@ tw:
     - name: SCSS
   vulnerabilityScanners:
     - "本專案為在地化成果，未涉及軟體開發作業。"
-  overrides:
-    repoOwners:
-      - "其他"
-  mainCopyrightOwnerLogo: "/images/projects/logos/gov-uk-logo.webp"
   # GitHub API 有資訊可以取得, 暫時手動填寫
   createdDate: "2023-02-03"
