+++ publiccode-parser/projects/gov-uk-notify.yml
@@ -45,8 +45,4 @@ tw:
   vulnerabilityScanners:
     - "本專案為在地化成果，未涉及軟體開發作業。"
   # GitHub API 有資訊可以取得, 暫時手動填寫
-  overrides:
-    repoOwners:
-      - "其他"
-  mainCopyrightOwnerLogo: "/images/projects/logos/gov-uk-logo.webp"
   createdDate: "2022-11-16"
