+++ publiccode-parser/projects/matrix-client-matrix-react-sdk.yml
@@ -44,9 +44,5 @@ tw:
     - "本專案為在地化成果，未涉及軟體開發作業。"
   techStacks:
     - name: "React"
-  mainCopyrightOwnerLogo: "/images/projects/logos/matrix-element-logo.webp"
-  overrides:
-    repoOwners:
-      - "其他"
   # GitHub API 有資訊可以取得, 暫時手動填寫
   createdDate: "2022-12-16"
