+++ publiccode-parser/projects/matrix-client-vector-im.element-web.yml
@@ -44,9 +44,5 @@ tw:
     - name: "React"
   vulnerabilityScanners:
     - "本專案為在地化成果，未涉及軟體開發"
-  overrides:
-    repoOwners:
-      - "其他"
-  mainCopyrightOwnerLogo: "/images/projects/logos/matrix-element-logo.webp"
   # GitHub API 有資訊可以取得, 暫時手動填寫
   createdDate: "2022-12-16"
