+++ publiccode-parser/projects/odf-application-tools.yml
@@ -45,6 +45,5 @@ tw:
   countryExtensionVersion: "1.0.0"
   vulnerabilityScanners:
     - "SonarQube"
-  mainCopyrightOwnerLogo: "/images/projects/logos/MODA-logo.svg"
   # GitHub API 有資訊可以取得, 暫時手動填寫
   createdDate: "2022-03-06"
