+++ publiccode-parser/projects/pdis-lighthouse.yml
@@ -42,5 +42,4 @@ tw:
   vulnerabilityScanners:
     - "GitHub CodeQL"
     - "GitHub Dependabot Alerts"
-  mainCopyrightOwnerLogo: "/images/projects/logos/PDIS-logo.webp"
   accessibilityConformance: null
