+++ publiccode-parser/projects/yivi-privacybydesign.irmamobile.yml
@@ -43,9 +43,5 @@ tw:
   countryExtensionVersion: "1.0.0"
   vulnerabilityScanners:
     - "本專案為在地化成果，未涉及軟體開發作業。"
-  overrides:
-    repoOwners:
-      - "其他"
-  mainCopyrightOwnerLogo: "/images/projects/logos/yivi-logo.webp"
   # GitHub API 有資訊可以取得, 暫時手動填寫
   createdDate: "2022-12-16"
